Passage token removed, func renamed
